Refactor tensor package name
